requirements.txt
pytest==7.3.1
selene==2.0.0rc4
iniconfig==2.0.0
python-dotenv==0.20.0
allure-pytest==2.13.2